+++ src/main/java/wooteco/subway/controller/LineController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import wooteco.subway.controller.converter.LineConverter;
 import wooteco.subway.dto.info.LineServiceRequest;
 import wooteco.subway.dto.info.LineServiceResponse;
 import wooteco.subway.dto.info.LineUpdateRequest;
