+++ src/main/java/wooteco/subway/controller/SectionController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import wooteco.subway.controller.converter.SectionConverter;
 import wooteco.subway.dto.request.SectionRequest;
 import wooteco.subway.service.SectionService;
