+++ src/main/java/wooteco/subway/controller/converter/LineConverter.java
@@ -1,4 +1,4 @@
-package wooteco.subway.controller;
+package wooteco.subway.controller.converter;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -11,16 +11,16 @@ import wooteco.subway.dto.response.LineResponse;
 import wooteco.subway.dto.response.StationResponse;
 
 public class LineConverter {
-    static LineUpdateRequest toInfo(Long id, LineRequest lineRequest) {
+    public static LineUpdateRequest toInfo(Long id, LineRequest lineRequest) {
         return new LineUpdateRequest(id, lineRequest.getName(), lineRequest.getColor(), lineRequest.getExtraFare());
     }
 
-    static LineServiceRequest toInfo(LineRequest lineRequest) {
+    public static LineServiceRequest toInfo(LineRequest lineRequest) {
         return new LineServiceRequest(lineRequest.getName(), lineRequest.getColor(), lineRequest.getUpStationId(),
             lineRequest.getDownStationId(), lineRequest.getDistance(), lineRequest.getExtraFare());
     }
 
-    static LineResponse toResponse(LineServiceResponse lineServiceResponse) {
+    public static LineResponse toResponse(LineServiceResponse lineServiceResponse) {
         List<StationResponse> stationResponses = lineServiceResponse.getStationInfos().stream()
             .map(StationConverter::toResponse)
             .collect(Collectors.toList());
