+++ src/main/java/wooteco/subway/controller/converter/SectionConverter.java
@@ -1,4 +1,4 @@
-package wooteco.subway.controller;
+package wooteco.subway.controller.converter;
 
 import wooteco.subway.dto.info.SectionCreateRequest;
 import wooteco.subway.dto.info.SectionDeleteRequest;
@@ -6,12 +6,12 @@ import wooteco.subway.dto.request.SectionRequest;
 
 public class SectionConverter {
 
-    static SectionCreateRequest toInfo(Long lineId, SectionRequest sectionRequest) {
+    public static SectionCreateRequest toInfo(Long lineId, SectionRequest sectionRequest) {
         return new SectionCreateRequest(lineId, sectionRequest.getUpStationId(), sectionRequest.getDownStationId(),
             sectionRequest.getDistance());
     }
 
-    static SectionDeleteRequest toInfo(Long lineId, Long stationId) {
+    public static SectionDeleteRequest toInfo(Long lineId, Long stationId) {
         return new SectionDeleteRequest(lineId, stationId);
     }
 }
